more content

+++ bubblesort/bubblesort.h
@@ -1,6 +1,6 @@
 #ifndef BUBBLESORT_H
 #define BUBBLESORT_H
 
-void bubblesort(int *a, int n);
+void bubblesort(int *a, size_t n);
 
 #endif
